# [chore][pkg/stanza] Speed up file deduplication in finder